+++ packages/types/CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.35.0 - 2024-09-23
+
 ### Added
 
 - Added NIP-11 `NostrRelayInfo` type.
 - NPolicy: added `info` getter that optionally returns a `NostrRelayInfo` object.
+- BREAKING: NRelay: added a required `.close()` method.
 
 ## 0.30.1 - 2024-09-09